new_waypoint_server/requirements.txt
alembic==1.7.7
aniso8601==9.0.1
click==8.0.4
Flask==2.0.3
Flask-API==3.0.post1
flask-blueprint==1.3.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
